+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 0.1.1 (2014-0X-XX)
 ------------------
 
+- Added: Python 2.6 support
+- Fixed: Resource representation was broken in Python 2.x, i.e. __repr__()
 - Fixed: dir() call on a resource object didn't work in Python 3.2.x
 - Changed: WikiPage resource refresh() method now automatically determines it's project_id